packages/capture/pyproject.toml
[project]
name = "openadapt-capture"
version = "0.1.0"
description = "GUI interaction capture - platform-agnostic event streams with time-aligned media"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "MLDSAI Inc.", email = "richard@mldsai.com"}
]
keywords = ["gui", "automation", "capture", "events", "rpa"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pydantic>=2.0.0",
    "pynput>=1.7.0",
    "av>=10.0.0",
    "Pillow>=9.0.0",
    "fire>=0.7.1",
    "mss>=6.0.0",
    "sounddevice>=0.5.3",
    "soundfile>=0.13.1",
    "openai-whisper>=20250625",
    "pydantic-settings>=2.12.0",
    "openai>=2.11.0",
]

[project.optional-dependencies]
# Whisper transcription (large dependency)
transcribe = [
    "openai-whisper>=20230314",
]

# Privacy scrubbing support
privacy = [
    "openadapt-privacy>=0.1.0",
]

# Everything
all = [
    "openadapt-capture[transcribe,privacy]",
]

dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "matplotlib>=3.5.0",
    "numpy>=1.21.0",
]

[project.scripts]
capture = "openadapt_capture.cli:main"

[project.urls]
Homepage = "https://github.com/OpenAdaptAI/openadapt-capture"
Repository = "https://github.com/OpenAdaptAI/openadapt-capture"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["openadapt_capture"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[dependency-groups]
dev = [
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
]
